Enable local startup on JDK 21
- Upgrade Lombok for JDK 21 compatibility
- Replace deprecated sun.misc.BASE64Encoder with java.util.Base64
- Add application-local profile (HTTP port 8080, no SSL)
- Move HTTPS redirect config to profile-gated TomcatHttpsRedirectConfig
