+++ pom.xml
@@ -16,6 +16,7 @@
 
     <properties>
         <java.version>1.8</java.version>
+        <lombok.version>1.18.34</lombok.version>
         <start-class>com.treeliked.darkme2.Darkme2Application</start-class>
     </properties>
